% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_clinical_trials.R
\name{count_clinical_trials}
\alias{count_clinical_trials}
\title{Find the No. of Trials in CT.gov website for a query}
\usage{
count_clinical_trials(query = NULL)
}
\arguments{
\item{query}{: string}
}
\value{
found_studies : integer, No. of Studies available in CT.gov database against a query.
}
\description{
Find the No. of Trials in CT.gov website for a query
}
\examples{
count_clinical_trials(query = "nsclc")

count_clinical_trials(query = "covid19 AND (new york OR italy)")

}
