% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTScraper-package.R
\docType{package}
\name{CTScraper-package}
\alias{CTScraper}
\alias{CTScraper-package}
\title{CTScraper: A Package for scraping CT.gov trials}
\description{
The Package scrapes CT.gov trials based on queries. The scraping is in realtime. 
  The return-type is a dataframe that can be used for future Analytics.
}
\author{
\strong{Maintainer}: Shahrukh Iqbal \email{shaiq681@student.liu.se} (shaiq681)

Authors:
\itemize{
  \item Mowniesh Asokan \email{mowas455@student.liu.se} (mowas455)
}

}
\keyword{internal}
